# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "get": {
        "operationId": "UserController_findAll",
        "summary": "List all users (paginated)",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UserController_findById",
        "summary": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/kyc": {
      "patch": {
        "operationId": "UserController_updateKyc",
        "summary": "Update user KYC status",
        "description": "Updates KYC verification status. Invalidates cached KYC eligibility.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKycDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "KYC status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/kyc-eligibility": {
      "get": {
        "operationId": "UserController_checkKycEligibility",
        "summary": "Check KYC eligibility for seat reservation",
        "description": "Checks if the user is eligible to reserve a seat based on their KYC status. Uses cached status (30s TTL).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "KYC eligibility result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycEligibilityResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/seats/jameya/{jameyaId}": {
      "get": {
        "operationId": "SeatController_findByJameya",
        "summary": "Get all seats for a Jameya",
        "description": "Returns all seats regardless of status, ordered by seat number.",
        "parameters": [
          {
            "name": "jameyaId",
            "required": true,
            "in": "path",
            "description": "Jameya UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of seats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatSummary"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "seats"
        ]
      }
    },
    "/api/seats/jameya/{jameyaId}/available": {
      "get": {
        "operationId": "SeatController_getAvailable",
        "summary": "Get available seats for a Jameya",
        "description": "Returns only seats with status AVAILABLE. Use this to show bookable seats. Note: slight staleness (~5s) is acceptable.",
        "parameters": [
          {
            "name": "jameyaId",
            "required": true,
            "in": "path",
            "description": "Jameya UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available seats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatSummary"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "seats"
        ]
      }
    },
    "/api/seats/{id}": {
      "get": {
        "operationId": "SeatController_findById",
        "summary": "Get seat by ID with Jameya details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Seat UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Seat details"
          },
          "404": {
            "description": "Seat not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "seats"
        ]
      }
    },
...
    "/api/reservations/user/{userId}": {
      "get": {
        "operationId": "ReservationController_findByUser",
        "summary": "Get all reservations for a user",
        "description": "Returns all reservations (PENDING, CONFIRMED, EXPIRED, CANCELLED) for a user, ordered by most recent.",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User reservations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationStatusResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reservations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Jameya Marketplace API",
    "description": "",
    "version": "1.0.0",
    "contact": {
      "name": "Jameya Marketplace Team",
      "url": "https://github.com/jameya-marketplace",
      "email": "team@jameya.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.jameya.io",
      "description": "Production (placeholder)"
    }
  ],
  "components": {
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "name": {
            "type": "string",
            "example": "Alice Johnson"
          },
          "phone": {
            "type": "string",
            "example": "+1234567001"
          },
          "kycStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED",
              "EXPIRED"
            ],
            "example": "VERIFIED"
          },
          "riskScore": {
            "type": "number",
            "example": 15,
            "description": "0-100 scale, lower = less risky"
          },
          "behaviorScore": {
            "type": "number",
            "example": 85,
            "description": "0-100 scale, higher = better"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "kycStatus",
          "riskScore",
          "behaviorScore",
          "createdAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "error": {
            "type": "string",
            "example": "ConflictException"
          },
          "message": {
            "type": "string",
            "example": "Seat is not available. Current status: RESERVED"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/reservations"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
      "UpdateKycDto": {
        "type": "object",
        "properties": {
          "kycStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED",
              "EXPIRED"
            ]
          }
        },
        "required": [
          "kycStatus"
        ]
      },
      "KycEligibilityResponse": {
        "type": "object",
        "properties": {
          "eligible": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED",
              "EXPIRED"
            ],
            "example": "VERIFIED"
          },
          "reason": {
            "type": "string",
            "example": "KYC verification is pending."
          }
        },
        "required": [
          "eligible",
          "status"
        ]
      },
...
      "SeatSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "seat-uuid-123"
          },
          "seatNumber": {
            "type": "number",
            "example": 1
          },
          "joiningPrice": {
            "type": "number",
            "example": 6000
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "CONFIRMED"
            ],
            "example": "AVAILABLE"
          }
        },
        "required": [
          "id",
          "seatNumber",
          "joiningPrice",
          "status"
        ]
      },
...
      "ReservationStatusResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "reservation-uuid-123"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-01-15T10:33:00.000Z"
          },
          "isExpired": {
            "type": "boolean",
            "example": false
          },
          "remainingSeconds": {
            "type": "number",
            "example": 145,
            "description": "Seconds remaining before expiry"
          }
        },
        "required": [
          "id",
          "status",
          "expiresAt",
          "isExpired",
          "remainingSeconds"
        ]
      },
...
    }
  }
}